:bug: NUMERIC->REAL

+++ v1/db/postgresql/migrations/2021-12-03-150805_fts_texts/up.sql
@@ -79,7 +79,7 @@ CREATE INDEX full_text_search_weighted__unaccent_idx
 
 CREATE OR REPLACE FUNCTION query_pali(query_str TEXT) 
   RETURNS TABLE(
-          rank NUMERIC,
+          rank REAL,
           paragraph INTEGER,
           wid VARCHAR,
           bold_single TEXT,
